did half of the k-tree
